// src/types/api.types.ts
  export enum ObservationCategory {
    PEST = 'PEST',
    DISEASE = 'DISEASE',
    BENEFICIAL = 'BENEFICIAL'
  }
  
  export enum SpeciesCode {
    // Pests
    THRIPS = 'THRIPS',
    RED_SPIDER_MITE = 'RED_SPIDER_MITE',
    WHITEFLIES = 'WHITEFLIES',
    MEALYBUGS = 'MEALYBUGS',
    CATERPILLARS = 'CATERPILLARS',
    FALSE_CODLING_MOTH = 'FALSE_CODLING_MOTH',
    PEST_OTHER = 'PEST_OTHER',
    
    // Diseases
    DOWNY_MILDEW = 'DOWNY_MILDEW',
    POWDERY_MILDEW = 'POWDERY_MILDEW',
    BOTRYTIS = 'BOTRYTIS',
    VERTICILLIUM = 'VERTICILLIUM',
    BACTERIAL_WILT = 'BACTERIAL_WILT',
    DISEASE_OTHER = 'DISEASE_OTHER',
    
    // Beneficial
    BENEFICIAL_PP = 'BENEFICIAL_PP'
  }